requirements/requirements.txt
# common 
numpy
pyyaml
rospkg
wind-pypcd
rich

# capture
PyGObject
opencv-python



